+++ UserDataDownloadSubOrchestrator/function.json
@@ -1,0 +1,10 @@
+{
+    "bindings": [
+      {
+        "name": "context",
+        "type": "orchestrationTrigger",
+        "direction": "in"
+      }
+    ],
+    "scriptFile": "../dist/UserDataDownloadSubOrchestrator/index.js"
+}
